refactor(shell): separate shell handler interface and local and SSH shell handler implementations
